Reorganize code to satisfy @tg44's nitpicks ;)
