+++ google-cloud-bigquery/src/main/scala/akka/stream/alpakka/googlecloud/bigquery/InsertAllRetryPolicy.scala
@@ -9,17 +9,21 @@ sealed abstract class InsertAllRetryPolicy {
   def deduplicate: Boolean
 }
 
-case object NoRetries extends InsertAllRetryPolicy {
-  override def retry: Boolean = false
-  override def deduplicate: Boolean = false
-}
+object InsertAllRetryPolicy {
 
-case object RetryNoDeduplication extends InsertAllRetryPolicy {
-  override def retry: Boolean = true
-  override def deduplicate: Boolean = false
-}
+  case object NoRetries extends InsertAllRetryPolicy {
+    override def retry: Boolean = false
+    override def deduplicate: Boolean = false
+  }
 
-case object RetryWithDeduplication extends InsertAllRetryPolicy {
-  override def retry: Boolean = true
-  override def deduplicate: Boolean = true
+  case object RetryNoDeduplication extends InsertAllRetryPolicy {
+    override def retry: Boolean = true
+    override def deduplicate: Boolean = false
+  }
+
+  case object RetryWithDeduplication extends InsertAllRetryPolicy {
+    override def retry: Boolean = true
+    override def deduplicate: Boolean = true
+  }
+
 }
